fix(benchmark): archive asyncio cancellation; run condition tests in CI
- The run-conditions tests lived in tests/test_community_bench.py, which
  is marked requires_mlx (skipped on the Linux lane) and is not in the
  Apple lane roster, so CI never executed them and changed-lines
  coverage failed. They move to an MLX-free tests/test_community_bench_conditions.py
  (Linux lane) which is also listed in the Apple lane so the in-process
  NSProcessInfo path runs on real hardware; a new test covers the
  Objective-C runtime failing to load.
- run_local names asyncio.CancelledError next to Exception so a cancelled
  benchmark still archives a cancelled record with its before snapshot
  (pr_validate codex finding).
